+++ ksyun/service_ksyun_vpc.go
@@ -466,6 +466,12 @@ func (s *VpcService) ReadSubnet(d *schema.ResourceData, subnetId string) (data m
 	}
 	for _, v := range results {
 		data = v.(map[string]interface{})
+
+		// 手动补AvailabilityZoneName字段，解决和配置字段不能对应引发的问题
+		// 例如: import的时候读不到这个值，导致认为这个值是空，plan会发现值变化
+		if v, ok := data["AvailabilityZoneName"]; ok {
+			data["AvailabilityZone"] = v
+		}
 	}
 	if len(data) == 0 {
 		return data, fmt.Errorf("Subnet %s not exist ", subnetId)
@@ -1459,7 +1465,7 @@ func (s *VpcService) CreateNetworkAclEntryWithAclCall(d *schema.ResourceData, r 
 			)
 			key := networkAclEntryHash(entry)
 			transform := make(map[string]SdkReqTransform)
-			for k, _ := range entry.(map[string]interface{}) {
+			for k := range entry.(map[string]interface{}) {
 				key := "network_acl_entries." + strconv.Itoa(key) + "." + k
 				if k == "icmp_type" || k == "icmp_code" {
 					transform[key] = SdkReqTransform{
@@ -1721,7 +1727,7 @@ func (s *VpcService) ModifyNetworkAclEntryWithAclCall(d *schema.ResourceData, r 
 				)
 				index := networkAclEntryHash(entry)
 				transform := make(map[string]SdkReqTransform)
-				for k, _ := range entry.(map[string]interface{}) {
+				for k := range entry.(map[string]interface{}) {
 					key := "network_acl_entries." + strconv.Itoa(index) + "." + k
 					transform[key] = SdkReqTransform{mapping: Downline2Hump(k)}
 				}
@@ -2115,7 +2121,7 @@ func (s *VpcService) CreateSecurityGroupEntryWithSgCall(d *schema.ResourceData, 
 			)
 			index := securityGroupEntryHash(entry)
 			transform := make(map[string]SdkReqTransform)
-			for k, _ := range entry.(map[string]interface{}) {
+			for k := range entry.(map[string]interface{}) {
 				key := "security_group_entries." + strconv.Itoa(index) + "." + k
 				if k == "icmp_type" || k == "icmp_code" {
 					transform[key] = SdkReqTransform{
@@ -2350,7 +2356,7 @@ func (s *VpcService) ModifySecurityGroupEntryWithSgCall(d *schema.ResourceData, 
 				)
 				index := securityGroupEntryHash(entry)
 				transform := make(map[string]SdkReqTransform)
-				for k, _ := range entry.(map[string]interface{}) {
+				for k := range entry.(map[string]interface{}) {
 					key := "security_group_entries." + strconv.Itoa(index) + "." + k
 					transform[key] = SdkReqTransform{mapping: Downline2Hump(k)}
 				}